+++ base/src/main/java/racecontrol/gui/app/livetiming/timing/tablemodels/StatsTableModel.java
@@ -30,6 +30,7 @@ import static racecontrol.client.extension.statistics.CarProperties.RACE_START_P
 import static racecontrol.client.extension.statistics.CarProperties.SPEED_TRAP_SPEED;
 import static racecontrol.gui.LookAndFeel.COLOR_PURPLE;
 import racecontrol.gui.app.livetiming.timing.tablemodels.columns.CarNumberColumn;
+import racecontrol.gui.app.livetiming.timing.tablemodels.columns.ConstructorColumn;
 import racecontrol.gui.app.livetiming.timing.tablemodels.columns.NameColumn;
 import racecontrol.gui.app.livetiming.timing.tablemodels.columns.PitFlagColumn;
 import racecontrol.gui.app.livetiming.timing.tablemodels.columns.PositionColumn;
@@ -47,8 +48,9 @@ public class StatsTableModel
         return new LPTableColumn[]{
             new PositionColumn(),
             new NameColumn(),
-            new PitFlagColumn(),
+            new ConstructorColumn(),
             new CarNumberColumn(),
+            new PitFlagColumn(),
             new LPTableColumn("+/-")
             .setMinWidth(100)
             .setCellRenderer((applet, context) -> placesLostGainedRenderer(applet, context)),